modularized table creation drop populate
