+++ main.js
@@ -12,41 +12,41 @@ const errorHandler = (err) => {
 
 
 
-const dropEmployees = () => {
-  db.run(`DROP TABLE employees`, (err)=>{
-    errorHandler(err);
-  });
-};
+// const dropEmployees = () => {
+//   db.run(`DROP TABLE employees`, (err)=>{
+//     errorHandler(err);
+//   });
+// };
 
-// dropEmployees();
+// // dropEmployees();
 
 //create the employee table if it does not exist
 //will not execute if the table exists
-db.run("CREATE TABLE IF NOT EXISTS employees(id INT, first TEXT, last TEXT, salary INT, dept TEXT, yearsWithCompany INT)", (err)=>{
-  errorHandler(err);
-});
+// db.run("CREATE TABLE IF NOT EXISTS employees(id INT, first TEXT, last TEXT, salary INT, dept TEXT, yearsWithCompany INT)", (err)=>{
+//   errorHandler(err);
+// });
 
  // db.run("INSERT INTO employees VALUES(1, 'Ashley', 'Irwin', 50000)");
 
- const populateEmployees = () => {
+ // const populateEmployees = () => {
 
-  const {list} = require('./employees.json');
+ //  const {list} = require('./employees.json');
 
-  list.forEach(each => {
-    db.run(`INSERT INTO employees VALUES(
-        ${each.id},
-        "${each.firstName}",
-        "${each.lastName}",
-        ${each.salary},
-        "${each.dept}",
-        ${each.yearsWithCompany}
-      )`, (err) =>{
-        errorHandler(err);
-      });
-  })
+ //  list.forEach(each => {
+ //    db.run(`INSERT INTO employees VALUES(
+ //        ${each.id},
+ //        "${each.firstName}",
+ //        "${each.lastName}",
+ //        ${each.salary},
+ //        "${each.dept}",
+ //        ${each.yearsWithCompany}
+ //      )`, (err) =>{
+ //        errorHandler(err);
+ //      });
+ //  })
 
- };
- // populateEmployees();
+ // };
+ // // populateEmployees();
 
 
 //db.get gets just the first row
